Make Xref_table an inner class of QPDF::Objects
